ISSUE #2149: Consolidated FindDistancePerMonth, FindElevationPerMonth and FindMovingTimePerMonth into FindTotalsPerMonth
