--- src/Domain/Rewind/FindElevationPerMonth/FindElevationPerMonth.php
@@ -1,24 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Rewind\FindElevationPerMonth;
-
-use App\Infrastructure\CQRS\Query\Query;
-use App\Infrastructure\ValueObject\Time\Years;
-
-/**
- * @implements Query<\App\Domain\Rewind\FindElevationPerMonth\FindElevationPerMonthResponse>
- */
-final readonly class FindElevationPerMonth implements Query
-{
-    public function __construct(
-        private Years $years,
-    ) {
-    }
-
-    public function getYears(): Years
-    {
-        return $this->years;
-    }
-}
