+++ src/Domain/Rewind/FindTotalsPerMonth/FindTotalsPerMonth.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Rewind\FindDistancePerMonth;
+namespace App\Domain\Rewind\FindTotalsPerMonth;
 
 use App\Infrastructure\CQRS\Query\Query;
 use App\Infrastructure\ValueObject\Time\Years;
 
 /**
- * @implements Query<\App\Domain\Rewind\FindDistancePerMonth\FindDistancePerMonthResponse>
+ * @implements Query<\App\Domain\Rewind\FindTotalsPerMonth\FindTotalsPerMonthResponse>
  */
-final readonly class FindDistancePerMonth implements Query
+final readonly class FindTotalsPerMonth implements Query
 {
     public function __construct(
         private Years $years,
