--- tests/Domain/Rewind/FindDistancePerMonth/FindDistancePerMonthQueryHandlerTest.php
@@ -1,90 +1,0 @@
-<?php
-
-namespace App\Tests\Domain\Rewind\FindDistancePerMonth;
-
-use App\Domain\Activity\ActivityId;
-use App\Domain\Activity\ActivityRepository;
-use App\Domain\Activity\ActivityWithRawData;
-use App\Domain\Activity\SportType\SportType;
-use App\Domain\Gear\GearId;
-use App\Domain\Rewind\FindDistancePerMonth\FindDistancePerMonth;
-use App\Domain\Rewind\FindDistancePerMonth\FindDistancePerMonthQueryHandler;
-use App\Infrastructure\ValueObject\Measurement\Length\Kilometer;
-use App\Infrastructure\ValueObject\Time\SerializableDateTime;
-use App\Infrastructure\ValueObject\Time\Year;
-use App\Infrastructure\ValueObject\Time\Years;
-use App\Tests\ContainerTestCase;
-use App\Tests\Domain\Activity\ActivityBuilder;
-
-class FindDistancePerMonthQueryHandlerTest extends ContainerTestCase
-{
-    private FindDistancePerMonthQueryHandler $queryHandler;
-
-    public function testHandle(): void
-    {
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('0'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-03-01 00:00:00'))
-                ->build(),
-            []
-        ));
-
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('1'))
-                ->withStartDateTime(SerializableDateTime::fromString('2025-01-01 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('2'))
-                ->withGearId(GearId::fromUnprefixed('3'))
-                ->withStartDateTime(SerializableDateTime::fromString('2023-01-01 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('3'))
-                ->withGearId(GearId::fromUnprefixed('2'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-01 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('4'))
-                ->withGearId(GearId::fromUnprefixed('5'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('8'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->build(),
-            []
-        ));
-
-        /** @var \App\Domain\Rewind\FindDistancePerMonth\FindDistancePerMonthResponse $response */
-        $response = $this->queryHandler->handle(new FindDistancePerMonth(Years::fromArray([Year::fromInt(2024)])));
-        $this->assertEquals(
-            [
-                [1, SportType::RIDE, Kilometer::from(30)],
-                [3, SportType::RIDE, Kilometer::from(10)],
-            ],
-            $response->getDistancePerMonth(),
-        );
-    }
-
-    #[\Override]
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->queryHandler = new FindDistancePerMonthQueryHandler($this->getConnection());
-    }
-}
