--- tests/Domain/Rewind/FindMovingTimePerMonth/FindMovingTimePerMonthQueryHandlerTest.php
@@ -1,96 +1,0 @@
-<?php
-
-namespace App\Tests\Domain\Rewind\FindMovingTimePerMonth;
-
-use App\Domain\Activity\ActivityId;
-use App\Domain\Activity\ActivityRepository;
-use App\Domain\Activity\ActivityWithRawData;
-use App\Domain\Activity\SportType\SportType;
-use App\Domain\Gear\GearId;
-use App\Domain\Rewind\FindMovingTimePerMonth\FindMovingTimePerMonth;
-use App\Domain\Rewind\FindMovingTimePerMonth\FindMovingTimePerMonthQueryHandler;
-use App\Infrastructure\ValueObject\Time\SerializableDateTime;
-use App\Infrastructure\ValueObject\Time\Year;
-use App\Infrastructure\ValueObject\Time\Years;
-use App\Tests\ContainerTestCase;
-use App\Tests\Domain\Activity\ActivityBuilder;
-
-class FindMovingTimePerMonthQueryHandlerTest extends ContainerTestCase
-{
-    private FindMovingTimePerMonthQueryHandler $queryHandler;
-
-    public function testHandle(): void
-    {
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('0'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-03-01 00:00:00'))
-                ->withMovingTimeInSeconds(100)
-                ->build(),
-            []
-        ));
-
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('1'))
-                ->withStartDateTime(SerializableDateTime::fromString('2025-01-01 00:00:00'))
-                ->withMovingTimeInSeconds(100)
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('2'))
-                ->withGearId(GearId::fromUnprefixed('3'))
-                ->withStartDateTime(SerializableDateTime::fromString('2023-01-01 00:00:00'))
-                ->withMovingTimeInSeconds(100)
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('3'))
-                ->withGearId(GearId::fromUnprefixed('2'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-01 00:00:00'))
-                ->withMovingTimeInSeconds(100)
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('4'))
-                ->withGearId(GearId::fromUnprefixed('5'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->withMovingTimeInSeconds(100)
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('8'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->withMovingTimeInSeconds(100)
-                ->build(),
-            []
-        ));
-
-        /** @var \App\Domain\Rewind\FindMovingTimePerMonth\FindMovingTimePerMonthResponse $response */
-        $response = $this->queryHandler->handle(new FindMovingTimePerMonth(Years::fromArray([Year::fromInt(2024)])));
-        $this->assertEquals(
-            [
-                [1, SportType::RIDE, 300],
-                [3, SportType::RIDE, 100],
-            ],
-            $response->getMovingTimePerMonth(),
-        );
-        $this->assertEquals(400, $response->getTotalMovingTime());
-    }
-
-    #[\Override]
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->queryHandler = new FindMovingTimePerMonthQueryHandler($this->getConnection());
-    }
-}
