+++ tests/Domain/Rewind/FindTotalsPerMonth/FindTotalsPerMonthQueryHandlerTest.php
@@ -1,14 +1,15 @@
 <?php
 
-namespace App\Tests\Domain\Rewind\FindElevationPerMonth;
+namespace App\Tests\Domain\Rewind\FindTotalsPerMonth;
 
 use App\Domain\Activity\ActivityId;
 use App\Domain\Activity\ActivityRepository;
 use App\Domain\Activity\ActivityWithRawData;
 use App\Domain\Activity\SportType\SportType;
 use App\Domain\Gear\GearId;
-use App\Domain\Rewind\FindElevationPerMonth\FindElevationPerMonth;
-use App\Domain\Rewind\FindElevationPerMonth\FindElevationPerMonthQueryHandler;
+use App\Domain\Rewind\FindTotalsPerMonth\FindTotalsPerMonth;
+use App\Domain\Rewind\FindTotalsPerMonth\FindTotalsPerMonthQueryHandler;
+use App\Infrastructure\ValueObject\Measurement\Length\Kilometer;
 use App\Infrastructure\ValueObject\Measurement\Length\Meter;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
 use App\Infrastructure\ValueObject\Time\Year;
@@ -16,9 +17,9 @@ use App\Infrastructure\ValueObject\Time\Years;
 use App\Tests\ContainerTestCase;
 use App\Tests\Domain\Activity\ActivityBuilder;
 
-class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
+class FindTotalsPerMonthQueryHandlerTest extends ContainerTestCase
 {
-    private FindElevationPerMonthQueryHandler $queryHandler;
+    private FindTotalsPerMonthQueryHandler $queryHandler;
 
     public function testHandle(): void
     {
@@ -26,7 +27,8 @@ class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
             ActivityBuilder::fromDefaults()
                 ->withActivityId(ActivityId::fromUnprefixed('0'))
                 ->withStartDateTime(SerializableDateTime::fromString('2024-03-01 00:00:00'))
-                ->withElevation(Meter::from(10))
+                ->withElevation(Meter::from(100))
+                ->withMovingTimeInSeconds(100)
                 ->build(),
             []
         ));
@@ -35,6 +37,8 @@ class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
             ActivityBuilder::fromDefaults()
                 ->withActivityId(ActivityId::fromUnprefixed('1'))
                 ->withStartDateTime(SerializableDateTime::fromString('2025-01-01 00:00:00'))
+                ->withElevation(Meter::from(100))
+                ->withMovingTimeInSeconds(100)
                 ->build(),
             []
         ));
@@ -43,6 +47,8 @@ class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
                 ->withActivityId(ActivityId::fromUnprefixed('2'))
                 ->withGearId(GearId::fromUnprefixed('3'))
                 ->withStartDateTime(SerializableDateTime::fromString('2023-01-01 00:00:00'))
+                ->withElevation(Meter::from(100))
+                ->withMovingTimeInSeconds(100)
                 ->build(),
             []
         ));
@@ -51,7 +57,8 @@ class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
                 ->withActivityId(ActivityId::fromUnprefixed('3'))
                 ->withGearId(GearId::fromUnprefixed('2'))
                 ->withStartDateTime(SerializableDateTime::fromString('2024-01-01 00:00:00'))
-                ->withElevation(Meter::from(20))
+                ->withElevation(Meter::from(100))
+                ->withMovingTimeInSeconds(100)
                 ->build(),
             []
         ));
@@ -60,7 +67,8 @@ class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
                 ->withActivityId(ActivityId::fromUnprefixed('4'))
                 ->withGearId(GearId::fromUnprefixed('5'))
                 ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->withElevation(Meter::from(10))
+                ->withElevation(Meter::from(100))
+                ->withMovingTimeInSeconds(100)
                 ->build(),
             []
         ));
@@ -68,20 +76,40 @@ class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
             ActivityBuilder::fromDefaults()
                 ->withActivityId(ActivityId::fromUnprefixed('8'))
                 ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->withElevation(Meter::from(10))
+                ->withElevation(Meter::from(100))
+                ->withMovingTimeInSeconds(100)
                 ->build(),
             []
         ));
 
-        /** @var \App\Domain\Rewind\FindElevationPerMonth\FindElevationPerMonthResponse $response */
-        $response = $this->queryHandler->handle(new FindElevationPerMonth(Years::fromArray([Year::fromInt(2024)])));
+        /** @var \App\Domain\Rewind\FindTotalsPerMonth\FindTotalsPerMonthResponse $response */
+        $response = $this->queryHandler->handle(new FindTotalsPerMonth(Years::fromArray([Year::fromInt(2024)])));
+
         $this->assertEquals(
             [
-                [1, SportType::RIDE, Meter::from(40)],
-                [3, SportType::RIDE, Meter::from(10)],
+                [1, SportType::RIDE, Kilometer::from(30)],
+                [3, SportType::RIDE, Kilometer::from(10)],
+            ],
+            $response->getDistancePerMonth(),
+        );
+        $this->assertEquals(
+            [
+                [1, SportType::RIDE, Meter::from(300)],
+                [3, SportType::RIDE, Meter::from(100)],
             ],
             $response->getElevationPerMonth(),
         );
+        $this->assertEquals(
+            [
+                [1, SportType::RIDE, 300],
+                [3, SportType::RIDE, 100],
+            ],
+            $response->getMovingTimePerMonth(),
+        );
+
+        $this->assertEquals(Kilometer::from(40), $response->getTotalDistance());
+        $this->assertEquals(Meter::from(400), $response->getTotalElevation());
+        $this->assertEquals(400, $response->getTotalMovingTime());
     }
 
     #[\Override]
@@ -89,6 +117,6 @@ class FindElevationPerMonthQueryHandlerTest extends ContainerTestCase
     {
         parent::setUp();
 
-        $this->queryHandler = new FindElevationPerMonthQueryHandler($this->getConnection());
+        $this->queryHandler = new FindTotalsPerMonthQueryHandler($this->getConnection());
     }
 }
